Add AC context (#430)
