+++ src/test/java/com/powsybl/openloadflow/OperationalLimitsTest.java
@@ -15,6 +15,7 @@ import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import com.powsybl.iidm.network.test.ThreeWindingsTransformerNetworkFactory;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.DenseMatrixFactory;
+import com.powsybl.openloadflow.ac.outerloop.AcLoadFlowContext;
 import com.powsybl.openloadflow.ac.outerloop.AcLoadFlowParameters;
 import com.powsybl.openloadflow.ac.outerloop.AcloadFlowEngine;
 import com.powsybl.openloadflow.network.*;
@@ -52,8 +53,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch1 = lfNetwork.getBranchById("NHV1_NHV2_1");
         assertTrue(branch1.getI1().eval() < getLimitValueFromAcceptableDuration(branch1, Integer.MAX_VALUE, Branch.Side.ONE, LimitType.CURRENT));
         LfBranch branch2 = lfNetwork.getBranchById("NHV1_NHV2_2");
@@ -86,8 +89,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch = lfNetwork.getBranchById("DL");
         assertEquals(0.626, branch.getI1().eval(), DELTA);
         assertTrue(Double.isNaN(branch.getI2().eval()));
@@ -105,8 +110,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch1 = lfNetwork.getBranchById("3WT_leg_1");
         assertEquals(1.144, branch1.getI1().eval(), DELTA);
         assertTrue(Double.isNaN(branch1.getI2().eval()));
@@ -123,8 +130,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch1 = lfNetwork.getBranchById("NHV1_NHV2_1");
         assertEquals(302.444, branch1.getP1().eval() * PerUnit.SB, 10E-3);
         LfBranch branch2 = lfNetwork.getBranchById("NHV1_NHV2_2");
@@ -141,8 +150,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch1 = lfNetwork.getBranchById("NHV1_NHV2_1");
         assertEquals(302.444, branch1.getP1().eval() * PerUnit.SB, 10E-3);
         LfBranch branch2 = lfNetwork.getBranchById("NHV1_NHV2_2");
@@ -160,8 +171,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch = lfNetwork.getBranchById("DL");
         assertEquals(54.815, branch.getP1().eval() * PerUnit.SB, 10E-3);
         assertTrue(Double.isNaN(branch.getP2().eval()));
@@ -175,8 +188,10 @@ class OperationalLimitsTest extends AbstractLoadFlowNetworkFactory {
         assertEquals(1, lfNetworks.size());
         LfNetwork lfNetwork = lfNetworks.get(0);
         AcLoadFlowParameters acParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), parameters, parametersExt, false, Reporter.NO_OP);
-        AcloadFlowEngine engine = new AcloadFlowEngine(lfNetwork, acParameters);
-        engine.run();
+        try (var context = new AcLoadFlowContext(lfNetwork, acParameters)) {
+            new AcloadFlowEngine(context)
+                    .run();
+        }
         LfBranch branch1 = lfNetwork.getBranchById("3WT_leg_1");
         assertEquals(116.251, branch1.getP1().eval() * PerUnit.SB, DELTA);
         assertTrue(Double.isNaN(branch1.getP2().eval()));
